Code for saving and loading notebooks
Converted over to using "storage location" and "repository" abstraction to record where a notebook is stored.

This will allow notebooks to be saved/loaded to the file system under Electron. It also allows notebooks to be saved/loaded to the backend for the Online version of DFN.

Added methods to save a notebook.