Removed unecessary debounced text update that was causing a problem.
Removed the debounced text update in the model.

This was causing an issue that resulted in a change in the UI causing a change in the model, causing the change to then be unecessarily applied back to the UI.

This was noticable because it was causing a cursor reset back to the start of the code cell while typing which was kind of annoying.