Got an almost working Electron build after jumping through multiple hoops.
For a start the Electron installer has to be built outside the workspace.
Otherwise it gets really confused about the multiple levels of node_modules.

Also node_modules has to be explicitly hoisted, have the symlinks taken out
and dedup all modules. Otherwise Electron-builder will put duplicate packages
into the builder which causes all sorts of issues.

I've hooked the Electron-builder process to disable its installation of
npm modules. I'm not strictly sure this is necessary in the end, but
it gives me more confidence that node_modules are the way I need them
to be if Electron-builder doesn't touch them.

Originally I tried to use "pnpm deploy" to export the Electron project
to a directory outside the workspace, but this results in an error
(some file system permissions error) and also the resulting project
has Electron-builder missing from it's node_modules - so I couldn't even
build the project after "deploying" it. Instead I've implemented my
own script to "export" the Electron project outside the workspace
directory and this works pretty well.

I also tried updating to the latest Electron-builder, but I don't think
this improved anything.