No longer deserializing the notebook in the evaluation engine.
I realised that the deserialization process in the
evaluation engine is unecessary.

I've updated it to work on the raw serialized data.

This will be a small performance boost, but more than that
it possible paves the way for merging the data model into the
view model and simplifying the architecture of the UI.