Refactor to place cell id with an "instance id"
Prior to conversion to the markdown notebook format I'm removing
unecessary data from the notebook format.

It seems unecessary to save the per-cell unique id to the notebook
file seeing as how it can be generated each time the notebook is
loaded.

It is still necessary to serialize cell id to JSON to transfer a
notebook to the evaluation engine. The EE needs cell ids to be able
to relate error messages and outputs to cells.